requirements.txt
Flask==3.0.2
gunicorn==21.2.0

# HTTP + retries
requests==2.31.0
urllib3==2.1.0

# YAML config support
pyyaml>=6.0

# Google Sheets config support
gspread>=5.10.0
oauth2client>=4.1.3

# Google Cloud clients
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
google-api-core==2.17.0

# Optional: ensures CSV parsing always works
google-cloud-core==2.4.1
